Create FilterAuthor component
