+++ src/components/Filter.js
@@ -1,45 +1,16 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React from 'react';
 import PropTypes from 'prop-types';
-import uniqid from 'uniqid';
-import { loadAuthors, getAuthors } from '../store/slicers/author';
-import { loadTitles, getTitles } from '../store/slicers/title';
 import FilterTitle from './FilterTitle';
+import FilterAuthor from './FilterAuthor';
 
-const Filter = ({ filterTitle, filterAuthor }) => {
-  const dispatch = useDispatch();
-  const titles = useSelector(getTitles);
-  const authors = useSelector(getAuthors);
-
-  useEffect(() => {
-    dispatch(loadTitles());
-  }, []);
-
-  useEffect(() => {
-    dispatch(loadAuthors());
-  }, []);
-
-  return (
-    <>
-      <FilterTitle filterTitle={filterTitle} />
-      <h1>Filter</h1>
-      <select onChange={(e) => filterTitle(e.target.value)}>
-        <option selected disabled>--select by Title--</option>
-        {
-          titles.map((t) => (<option key={uniqid()} value={t}>{t}</option>))
-         }
-      </select>
-
-      <select onChange={(e) => filterAuthor(e.target.value)}>
-        <option selected disabled>--select by Author--</option>
-        {
-          authors.map((a) => (<option key={uniqid()} value={a}>{a}</option>))
-        }
-      </select>
-      <hr />
-    </>
-  );
-};
+const Filter = ({ filterTitle, filterAuthor }) => (
+  <>
+    <h1>Filter</h1>
+    <FilterTitle filterTitle={filterTitle} />
+    <FilterAuthor filterAuthor={filterAuthor} />
+    <hr />
+  </>
+);
 
 Filter.propTypes = {
   filterTitle: PropTypes.func.isRequired,
